requirements.txt
Flask
Flask-Cors
gunicorn
pandas
numpy
scikit-learn
scikit-surprise
gdown
